Organización del código, se mueve displayQuestions a un archivo separado questionDisplay.js
